fix database bad type

+++ OpenNos.Data/CharacterDTO.cs
@@ -34,7 +34,7 @@ namespace OpenNos.Data
         public int Mp { get; set; }
         public int ArenaWinner { get; set; }
         public long Reput { get; set; }
-        public byte Dignite { get; set; }
+        public short Dignite { get; set; }
         public long Gold { get; set; }
         public int Backpack { get; set; }
         public byte Level { get; set; }
